token authentication check for apis

+++ src/main/java/com/alikmndlu/userservice/controller/UserController.java
@@ -4,9 +4,12 @@ import com.alikmndlu.userservice.config.UrlConfig;
 import com.alikmndlu.userservice.dto.LoginCredentialsDto;
 import com.alikmndlu.userservice.dto.RegisterCredentialsDto;
 import com.alikmndlu.userservice.dto.UserAddressesDto;
+import com.alikmndlu.userservice.dto.UserIdEmailAddressDto;
 import com.alikmndlu.userservice.model.User;
 import com.alikmndlu.userservice.service.UserService;
+import com.alikmndlu.userservice.util.JwtUtil;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.RestTemplate;
@@ -18,9 +21,8 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/api/users")
 @RequiredArgsConstructor
+@Slf4j
 public class UserController {
-
-    private final RestTemplate restTemplate;
 
     private final UserService userService;
 
@@ -28,8 +30,18 @@ public class UserController {
 
     private final String apiHost = "http://localhost:9001/api/users";
 
+    private final JwtUtil jwtUtil;
+
     @GetMapping("/{user-id}")
-    public ResponseEntity<User> findUser(@PathVariable("user-id") Long userId) {
+    public ResponseEntity<User> findUser(
+            @PathVariable("user-id") Long userId,
+            @RequestHeader("Authorization") String token) {
+
+        UserIdEmailAddressDto userInfo = jwtUtil.transferTokenToIdAndEmailAddress(token);
+        if (!userInfo.getId().equals(userId)) {
+            return ResponseEntity.badRequest().build();
+        }
+
         Optional<User> user = userService.findUserById(userId);
 
         return user.isPresent() ?
@@ -42,16 +54,37 @@ public class UserController {
         return ResponseEntity.accepted().body(userService.updateUser(userId, updates));
     }
 
+    // Delete User By Id
     @DeleteMapping("/{user-id}")
-    public void deleteUser(@PathVariable("user-id") Long userId) {
+    public ResponseEntity<?> deleteUser(
+            @PathVariable("user-id") Long userId,
+            @RequestHeader("Authorization") String token) {
+
+        UserIdEmailAddressDto userInfo = jwtUtil.transferTokenToIdAndEmailAddress(token);
+        log.info("Delete User API {requestId: {},jwtId: {}, jwtEmailAddress: {}, result: {}}", userId, userInfo.getId(), userInfo.getEmailAddress(), userInfo.getId().equals(userId));
+        if (!userInfo.getId().equals(userId)) {
+            return ResponseEntity.badRequest().build();
+        }
+
         userService.deleteUser(userId);
+        return ResponseEntity.ok().build();
     }
 
+    // Get USer With Addresses List
     @GetMapping("/addresses/{user-id}")
-    public ResponseEntity<UserAddressesDto> findUserWithAddresses(@PathVariable("user-id") Long userId) {
+    public ResponseEntity<UserAddressesDto> findUserWithAddresses(
+            @PathVariable("user-id") Long userId,
+            @RequestHeader("Authorization") String token) {
+
+        UserIdEmailAddressDto userInfo = jwtUtil.transferTokenToIdAndEmailAddress(token);
+        log.info("Delete User API {requestId: {},jwtId: {}, jwtEmailAddress: {}, result: {}}", userId, userInfo.getId(), userInfo.getEmailAddress(), userInfo.getId().equals(userId));
+        if (!userInfo.getId().equals(userId)) {
+            return ResponseEntity.badRequest().build();
+        }
+
         Optional<User> user = userService.findUserById(userId);
 
-        if (user.isEmpty()) {
+        if (user.isPresent()) {
             return ResponseEntity.notFound().build();
         } else {
             return ResponseEntity.ok().body(userService.findUserWithAddresses(user.get()));
@@ -76,6 +109,14 @@ public class UserController {
                         .password(registerDto.getPassword())
                         .build()
         );
+        return ResponseEntity.ok().body(user);
+    }
+
+    // Find User By Email Address
+    @GetMapping("/by-email/{email-address}")
+    public ResponseEntity<User> findUserByEmailAddress(@PathVariable("email-address") String emailAddress) {
+        User user = userService.findByEmailAddress(emailAddress);
+        System.out.println("Request is Here");
         return ResponseEntity.ok().body(user);
     }
 }
